refactor game logic
move all game logic modules into one file together, adjust links to these files in the test files, add method in gameboard that returns the gameboard
