+++ battleship/src/app/game-logic/ship.js
